fix: moved filterItem to separate file
